# api-definitions/ai-openapi.json
{
  "components": {
    "schemas": {
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressBySearchResponse": {
        "properties": {
          "recommendationCode": {
            "type": "string"
          },
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseAddressNonIDS"
            }
          },
          "offset": {
            "type": "integer",
            "format": "int32"
          },
          "latitude": {
            "type": "string"
          },
          "historical": {
            "type": "boolean"
          },
          "epoch": {
            "type": "string"
          },
          "sampleSize": {
            "type": "integer",
            "format": "int64"
          },
          "maxScore": {
            "type": "number",
            "format": "double"
          },
          "verbose": {
            "type": "boolean"
          },
          "filter": {
            "type": "string"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "rangekm": {
            "type": "string"
          },
          "pafdefault": {
            "type": "boolean"
          },
          "countryBoosts": {
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.CountryBoosts"
          },
          "limit": {
            "type": "integer",
            "format": "int32"
          },
          "matchtype": {
            "type": "string"
          },
          "tokens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Token"
            }
          },
          "matchthreshold": {
            "type": "number",
            "format": "float"
          },
          "longitude": {
            "type": "string"
          }
        },
        "required": [
          "tokens",
          "addresses",
          "matchtype",
          "recommendationCode",
          "filter",
          "historical",
          "epoch",
          "rangekm",
          "latitude",
          "longitude",
          "limit",
          "offset",
          "total",
          "sampleSize",
          "maxScore",
          "matchthreshold",
          "verbose",
          "countryBoosts",
          "pafdefault"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseGeo": {
        "properties": {
          "latitude": {
            "type": "number",
            "format": "decimal"
          },
          "longitude": {
            "type": "number",
            "format": "decimal"
          },
          "easting": {
            "type": "integer",
            "format": "int32"
          },
          "northing": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "latitude",
          "longitude",
          "easting",
          "northing"
        ]
      },
      "Token": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          },
          "label": {
            "type": "string",
            "enum": [
              "OrganisationName",
              "DepartmentName",
              "SubBuildingName",
              "BuildingName",
              "BuildingNumber",
              "StreetName",
              "Locality",
              "TownName",
              "Postcode"
            ]
          }
        }
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseRelative": {
        "properties": {
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "siblings": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "parents": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        },
        "required": [
          "level",
          "siblings",
          "parents"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseHighlightHit": {
        "properties": {
          "source": {
            "type": "string"
          },
          "lang": {
            "type": "string"
          },
          "distinctHitCount": {
            "type": "integer",
            "format": "int32"
          },
          "highLightedText": {
            "type": "string"
          }
        },
        "required": [
          "source",
          "lang",
          "distinctHitCount",
          "highLightedText"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressBySearchResponseContainer": {
        "properties": {
          "apiVersion": {
            "type": "string"
          },
          "dataVersion": {
            "type": "string"
          },
          "response": {
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressBySearchResponse"
          },
          "status": {
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseStatus"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseError"
            }
          }
        },
        "required": [
          "apiVersion",
          "dataVersion",
          "response",
          "status",
          "errors"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseCrossRef": {
        "properties": {
          "crossReference": {
            "type": "string"
          },
          "source": {
            "type": "string"
          }
        },
        "required": [
          "crossReference",
          "source"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseAddressNonIDS": {
        "properties": {
          "uprn": {
            "type": "string"
          },
          "parentUprn": {
            "type": "string"
          },
          "relatives": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseRelative"
            }
          },
          "crossRefs": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseCrossRef"
            }
          },
          "formattedAddress": {
            "type": "string"
          },
          "formattedAddressNag": {
            "type": "string"
          },
          "formattedAddressPaf": {
            "type": "string"
          },
          "welshFormattedAddressNag": {
            "type": "string"
          },
          "welshFormattedAddressPaf": {
            "type": "string"
          },
          "highlights": {
            "nullable": true,
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseHighlight"
          },
          "paf": {
            "nullable": true,
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponsePaf"
          },
          "nag": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseNag"
            }
          },
          "geo": {
            "nullable": true,
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseGeo"
          },
          "classificationCode": {
            "type": "string"
          },
          "countryCode": {
            "type": "string"
          },
          "lpiLogicalStatus": {
            "type": "string"
          },
          "confidenceScore": {
            "type": "number",
            "format": "double"
          },
          "underlyingScore": {
            "type": "number",
            "format": "float"
          }
        },
        "required": [
          "uprn",
          "parentUprn",
          "formattedAddress",
          "formattedAddressNag",
          "formattedAddressPaf",
          "welshFormattedAddressNag",
          "welshFormattedAddressPaf",
          "classificationCode",
          "countryCode",
          "lpiLogicalStatus",
          "confidenceScore",
          "underlyingScore"
        ]
      },
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseHighlight": {
        "properties": {
          "bestMatchAddress": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "lang": {
            "type": "string"
          },
          "hits": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseHighlightHit"
            }
          }
        },
        "required": [
          "bestMatchAddress",
          "source",
          "lang"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseError": {
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseStatus": {
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.CountryBoosts": {
        "properties": {
          "eboost": {
            "type": "number",
            "format": "double"
          },
          "nboost": {
            "type": "number",
            "format": "double"
          },
          "sboost": {
            "type": "number",
            "format": "double"
          },
          "wboost": {
            "type": "number",
            "format": "double"
          },
          "lboost": {
            "type": "number",
            "format": "double"
          },
          "mboost": {
            "type": "number",
            "format": "double"
          },
          "jboost": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "eboost",
          "nboost",
          "sboost",
          "wboost",
          "lboost",
          "mboost",
          "jboost"
        ]
      },
...
    }
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/addresses": {
      "get": {
        "operationId": "addressQuery",
        "tags": [
          "routes"
        ],
        "summary": "Search for an address.",
        "description": "Returns a ranked list of addresses matching the search query in the specified format.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "input",
            "description": "Specifies the address search string (e.g. 14 Acacia Avenue, Ruislip, HA4 8RG). Can only be blank if rangekm, lat, lon and classificationfilter are supplied in which case the underlyingScore returned is the distance in km.",
            "required": true
          },
          {
            "schema": {
              "default": 0,
              "nullable": true,
              "maximum": 250,
              "type": "integer"
            },
            "in": "query",
            "name": "offset",
            "description": "Specifies the offset from zero, used for pagination.",
            "required": false
          },
          {
            "schema": {
              "default": 10,
              "nullable": true,
              "maximum": 100,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "description": "Specifies the number of addresses to return.",
            "required": false
          },
          {
            "schema": {
              "nullable": true,
              "type": "string"
            },
            "in": "query",
            "name": "classificationfilter",
            "description": "Classification code filter. Can be pattern match (ZW*), exact match (RD06), multiple exact match (RD02,RD04) or a preset keyword such as residential, commercial or workplace",
            "required": false
          },
          {
            "schema": {
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "rangekm",
            "description": "Limit results to those within this number of kilometers of point (decimal e.g. 0.1)",
            "required": false
          },
          {
            "schema": {
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "lat",
            "description": "Latitude of point in decimal format (e.g. 50.705948).",
            "required": false
          },
          {
            "schema": {
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "lon",
            "description": "Longitude of point in decimal format (e.g. -3.5091076).",
            "required": false
          },
          {
            "schema": {
              "default": true,
              "nullable": true,
              "type": "boolean"
            },
            "in": "query",
            "name": "historical",
            "description": "Include historical addresses",
            "required": false
          },
          {
            "schema": {
              "default": 5,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "matchthreshold",
            "description": "Minimum confidence score (percentage) for match to be included in results.",
            "required": false
          },
          {
            "schema": {
              "default": false,
              "nullable": true,
              "type": "boolean"
            },
            "in": "query",
            "name": "verbose",
            "description": "Include the full address details in the response (including relatives, crossRefs, paf and nag).",
            "required": false
          },
          {
            "schema": {
              "nullable": true,
              "type": "string"
            },
            "in": "query",
            "name": "epoch",
            "description": "Select a specific AddressBase Epoch to search.",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "eboost",
            "description": "Set to 0 to exclude addresses in England",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "nboost",
            "description": "Set to 0 to exclude addresses in Northern Ireland",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "sboost",
            "description": "Set to 0 to exclude addresses in Scotland",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "wboost",
            "description": "Set to 0 to exclude addresses in Wales",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "lboost",
            "description": "Set to 0 to exclude addresses in Channel Islands",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "mboost",
            "description": "Set to 0 to exclude addresses in Isle of Man",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "jboost",
            "description": "Set to 0 to exclude addresses outside country boundaries",
            "required": false
          },
          {
            "schema": {
              "default": false,
              "nullable": true,
              "type": "boolean"
            },
            "in": "query",
            "name": "pafdefault",
            "description": "Populate the formatted address field with PAF if present",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Success. A json return of matched addresses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressBySearchResponseContainer"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. Indicates an issue with the request. Further details are provided in the response."
          },
          "401": {
            "description": "Unauthorised. The API key provided with the request is invalid."
          },
          "429": {
            "description": "Server too busy. The Address Index API is experiencing exceptional load."
          },
          "500": {
            "description": "Internal server error. Failed to process the request due to an internal error."
          }
        }
      }
    },
...
  },
  "info": {
    "title": "ONS Address Index API",
    "description": "The ONS Address Index API lets you retrieve the official version of addresses in England and Wales. The data is currently updated on a 6 week rolling basis.",
    "version": "see /version below"
  },
...
}